requirements.txt
pandas
matplotlib
nltk
scikit-learn
transformers
torch
sumy
diffusers
pillow
accelerate